extracting profile id
